+++ src/main/java/com/example/HealthCare/Services/PatientService.java
@@ -1,19 +1,15 @@
 package com.example.HealthCare.Services;
 
 
-import com.example.HealthCare.DTO.CreatePatientDTO;
-import com.example.HealthCare.DTO.PatientDTO;
+import com.example.HealthCare.DTO.PatientRequestDTO;
+import com.example.HealthCare.DTO.PatientResponseDTO;
 import com.example.HealthCare.Mapper.PatientMapper;
 import com.example.HealthCare.Models.Patient;
 import com.example.HealthCare.Repositories.PatientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.repository.core.support.RepositoryMethodInvocationListener;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class PatientService {
@@ -26,24 +22,25 @@ public class PatientService {
 
 
 
-   public void ajouterPatient(CreatePatientDTO patient){
-        patientRepository.save(patientMapper.toEntity(patient));
+   public PatientResponseDTO ajouterPatient(PatientRequestDTO patient){
+     Patient patient1 =  patientMapper.toEntity(patient);
+     return patientMapper.toDto(patientRepository.save(patient1));
     }
 
-    public List<PatientDTO> listerPatients(){
+    public List<PatientResponseDTO> listerPatients(){
 
             List<Patient> patients = patientRepository.findAll();
             return patients
                     .stream()
                     .map((patient)->{
-                        PatientDTO dto = patientMapper.toDto(patient);
+                        PatientResponseDTO dto = patientMapper.toDto(patient);
                         dto.setTotalRendezVous(patient.getRenderVousList().size());
                         return dto;
                     })
                     .toList();
     }
 
-    public Patient modifierPatient(int patientId, CreatePatientDTO newPatient){
+    public Patient modifierPatient(int patientId, PatientRequestDTO newPatient){
        Patient patient = patientRepository.findById(patientId).orElseThrow(()->new RuntimeException("Patient Not Found with id " + patientId));
        if(patient != null){
            patient.setNom(newPatient.getNom());
@@ -57,13 +54,18 @@ public class PatientService {
     }
 
 
-    public PatientDTO consulterPatient(int patientId){
+    public PatientResponseDTO consulterPatient(int patientId){
        Patient patient = patientRepository.findById(patientId).orElseThrow(()-> new RuntimeException("Patient Not Found !!"));
        return patientMapper.toDto(patient);
     }
 
-    public void supprimerPatient(int patientId){
-        patientRepository.deleteById(patientId);
+    public Boolean supprimerPatient(int patientId){
+        Patient patient = patientRepository.findById(patientId).orElseThrow(()-> new RuntimeException("Patient Not Found !!"));
+        if(patient != null){
+            patientRepository.deleteById(patientId);
+            return true;
+        }
+        return false;
     }
